Update java-stop.sh
added kill execution on condition

+++ java-stop.sh
@@ -1,2 +1,4 @@
 #!/usr/bin/env bash
-pkill -F /opt/hello/hello.pid
+if [ -f /opt/hello/hello.pid ]; then
+    pkill -F /opt/hello/hello.pid
+fi
